エラーハンドリングの追加

+++ app/controllers/image_controller.rb
@@ -2,8 +2,14 @@ class ImageController < ApplicationController
 
   def destroy
     image = Image.find(params[:id])
-    image.destroy
-    redirect_to :back
+    if image.product.seller_id == current_user.id
+      image.destroy
+      redirect_to :back
+      # flash[:alert] = '商品を削除しました'
+    else
+      redirect_to edit_products_path(image.product)
+      # flash[:alert] = '商品削除に失敗しました'
+    end
   end
